pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "antigravity-sdk"
version = "1.0.0"
description = "Python SDK for Antigravity Claude Proxy - Use Claude/Gemini in your apps with retry, tools, and conversations"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "DecentralizedJM", email = "decentralizedjm@gmail.com"}
]
keywords = [
    "claude",
    "anthropic", 
    "ai",
    "sdk",
    "telegram",
    "gemini",
    "llm",
    "chatbot",
    "antigravity",
    "proxy",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Communications :: Chat",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

dependencies = [
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
telegram = [
    "python-telegram-bot>=20.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "respx>=0.20.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "black>=23.0.0",
]
all = [
    "antigravity-sdk[telegram,dev]",
]

[project.urls]
"Homepage" = "https://github.com/DecentralizedJM/antigravity-proxy-claude-sdk"
"Documentation" = "https://github.com/DecentralizedJM/antigravity-proxy-claude-sdk#readme"
"Bug Tracker" = "https://github.com/DecentralizedJM/antigravity-proxy-claude-sdk/issues"
"Source" = "https://github.com/DecentralizedJM/antigravity-proxy-claude-sdk"
"Changelog" = "https://github.com/DecentralizedJM/antigravity-proxy-claude-sdk/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
antigravity_sdk = ["py.typed"]

# ============================================================================
# Testing
# ============================================================================

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.coverage.run]
source = ["src/antigravity_sdk"]
branch = true
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]

# ============================================================================
# Code Quality
# ============================================================================

[tool.ruff]
target-version = "py39"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "B905",  # zip strict
]

[tool.ruff.lint.isort]
known-first-party = ["antigravity_sdk"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
